requirements.txt
git+https://github.com/wlanslovenija/django.git@ab42fa765fe3c52bac4f0bdf8c2de48ad5044a3a#egg=Django-wlanslovenija
GDAL==1.10.1
Jinja2==2.7.2
MarkupSafe==0.23
Pygments==1.6
South==0.8.4
Sphinx==1.2.2
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.11
git+https://github.com/wlanslovenija/datastream.git@0b95d9495d751ec42a1759c4f234a7b580faf50d#egg=datastream-master
defusedxml==0.4.1
django-celery==3.1.10
django-classy-tags==0.5.1
git+https://github.com/wlanslovenija/django-datastream.git@79347949e54ae1ff2e97a05f6589b957bae2eda4#egg=django_datastream-master
django-guardian==1.2
git+https://github.com/mitar/django-missing.git@9cf0fd6cfe39bb2cd0b90b3dd37be1b465308ae8#egg=django_missing-master
git+https://github.com/wlanslovenija/django-overextends.git@ee963e2c338c1134e6ec49acabe9822b3636b536#egg=django_overextends-wlanslovenija
django-phonenumber-field==0.5
git+https://github.com/wlanslovenija/django_polymorphic.git@514f6079696838f241bec0c0397d8ce193c19ac0#egg=django_polymorphic-wlanslovenija
git+https://github.com/wlanslovenija/django-registration.git@670779dce3b81fabc7cc2394db1ddd45ffcc4dc7#egg=django_registration-wlanslovenija
django-sekizai==0.7
git+https://github.com/wlanslovenija/django-tastypie.git@493359475f5ef21ba071bb004699293e2675beed#egg=django_tastypie-wlanslovenija
dnspython==1.11.1
docutils==0.11
lxml==3.3.5
kombu==3.0.16
mimeparse==0.1.3
mock==1.0.1
mongoengine==0.8.1
phonenumbers==6.1.0
psycopg2==2.5.3
git+https://github.com/wlanslovenija/pyScss.git@b84cf224eec73df01241a4c7fecc064dcafd6c4d#egg=pyScss-wlanslovenija
pymongo==2.6.3
python-dateutil==2.2
python-memcached==1.53
pytz==2014.3
six==1.6.1
sphinx-me==0.2.1
ujson==1.33
wsgiref==0.1.2
django-timezone-field==1.0
django-json-field==0.5.5
django-uuidfield==0.5.0
requests==2.2.1
paramiko==1.15.1
